feat: add support for custom ID formatting globally
